Move the Event scopes to class methods
Here is the deal. The scopes defined with the `ActiveRecord::Base.scope`
macro are expected to return either `ActiveRecord::Relation` or a falsy
value. If the result is falsy (to be hones, most-likely nil) the scope
will return the previous one, in our case `Event.all`.

Given that Event.upcoming is defined as follows:

```ruby
class Event < ApplicationRecord
  scope :upcoming, -> {
    includes(location: :venue).order(time: :desc).where('time >= ?', Time.current).first
  }
end

Check what curious result we get in IRB:

```ruby
>> Event.upcoming # 👈 Defined by ActiveRecord::Relation
  Event Load (1.3ms)  SELECT  "events".* FROM "events" WHERE (time >= '2018-05-04 15:35:28.296722') ORDER BY "events"."time" DESC LIMIT $1  [["LIMIT", 1]]
  Event Load (0.4ms)  SELECT  "events".* FROM "events" LIMIT $1  [["LIMIT", 11]]
=> #<ActiveRecord::Relation [#<Event id: 1, time: "2017-02-27 05:00:00", description: "This will be our very first Ruby Banitsa, so hello...", created_at: "2017-05-05 11:38:01", updated_at: "2017-11-20 09:13:25", published_at: nil>, #<Event id: 2, time: "2017-03-15 14:30:00", description: "The second Ruby Banitsa. Exciting! Join us for som...", created_at: "2017-05-05 12:02:51", updated_at: "2017-11-20 09:13:07", published_at: nil>, #<Event id: 3, time: "2017-04-01 05:00:00", description: "Third time is the charm! Come listen to Radoslav S...", created_at: "2017-05-05 12:07:10", updated_at: "2017-11-20 09:11:48", published_at: nil>, #<Event id: 4, time: "2017-05-10 12:30:00", description: "After all the National Holidays in the world, let'...", created_at: "2017-05-05 12:09:59", updated_at: "2017-11-20 09:12:27", published_at: nil>, #<Event id: 5, time: "2017-11-20 15:30:00", description: "After a long summer break, we're back, friends! Th...", created_at: "2017-11-14 10:48:49", updated_at: "2017-11-20 08:51:33", published_at: nil>, #<Event id: 6, time: "2017-12-12 15:30:00", description: "The last Banitsa was great! We reconciled old frie...", created_at: "2017-12-01 12:25:36", updated_at: "2017-12-03 07:14:54", published_at: nil>, #<Event id: 7, time: "2018-01-22 15:30:00", description: "New year, new banitsa. This time, with fortunes. F...", created_at: "2018-01-04 08:59:11", updated_at: "2018-01-04 12:11:40", published_at: "2018-01-04 12:11:39">, #<Event id: 15, time: "2018-04-26 13:20:01", description: "asdasdasdasd", created_at: "2018-04-02 13:20:54", updated_at: "2018-04-02 13:20:54", published_at: nil>]>
```

My expectation is for `Event.upcoming` is to return `Event | nil`.
However, the reality is that when `nil` is returned, even if this
happening to be the result of another query, the value of `Event.all` is
return from the scope. This won't issue another query, until something
evaluates it, in this case, `#inspect` being called from IRB.

It's a really confusing problem to get the first time you hit it. 🤷
That's why, folks, it's better if you use class methods for them pesky
scopes! 😂