Stop using Location model
The `locations` table doesn't seem to be necessary. It's acting as a
middle table between `events` and `venues`. That's kind of confusing
because we expect a venue to have multiple events (has_many) and event
to be associated with a single venue (belongs_to).

In this commit we:

- Change the association between `events` and `venues` to one-to-many.
- Populate the newly added `events.venue_id` column.
- Remove any references to `Location` model.
- Remove the redundant `Event.create_with_venue`.
- Remove custom create action for events in the admin panel.
- More cleanup.

As a follow-up we'll:

- Drop `locations` table.
- Add `not null` constraint on `events.venue_id`.