Add NOT NULL constraint to events.venue_id
Once we ensure all events in the databaase are associated with venues,
we can improve the data integrity by making the `events.venue_id` column
not nullable.

The migration is done in a naive way since we have around 50 events at
the moment and blocking the table shouldn't be a problem.