Fix "Post a job" double hover animation
The "Jobs" sidebar is always visible and always in the same position. On
desktop browsers, if we click the button, the element gets re-rendered,
and the hover animation plays again, because the cursor ends up on the
same place in the new page with a new sidebar that looks exactly as the
previous one and a new hover animation plays on the new button.

If we persist the sidebar across turbo reloads we avoid the second
animation.