Force Active Storage to generate paths instead of inferred URLs
Active Storage internal controllers guess the protocol, host and port of
the URLs it generates from the request. Our server and deploy process
is not setup in away to forward them from the "real" domain, so we end
up with URLs like: http://localhost:3011/...

Not good. And we have to hack Active Storage to change this URL
inference. Not good. Again.